+++ src/CirculationLogList/CirculationLogListActions/useCirculationLogExport.test.js
@@ -31,12 +31,15 @@ const files = ['bursar1.dat', 'bursar2'];
 describe('useCirculationLogExport', () => {
   let postMock;
   let getMock;
+  let blobMock;
 
   beforeEach(() => {
+    blobMock = jest.fn().mockImplementation(() => Promise.resolve());
     postMock = jest.fn(() => ({
       json: jest.fn(() => ({ id: 'sas-fsd-53-sda' })),
     }));
     getMock = jest.fn(() => ({
+      blob: blobMock,
       json: jest.fn(() => ({ id: 'sas-fsd-53-sda', status: 'SUCCESSFUL', files })),
     }));
 
@@ -90,8 +93,6 @@ describe('useCirculationLogExport', () => {
   });
 
   it('should download all files when job is successful', async () => {
-    const linkMock = document.createElement('a');
-
     jest.clearAllTimers();
 
     const { result } = renderHook(
@@ -101,19 +102,14 @@ describe('useCirculationLogExport', () => {
       { wrapper },
     );
 
-    linkMock.dispatchEvent = jest.fn();
-    jest.spyOn(document, 'createElement').mockReturnValue(linkMock);
-
     await result.current.requestExport();
 
     jest.runAllTimers();
 
-    await waitFor(() => expect(linkMock.dispatchEvent).toHaveBeenCalledTimes(files.length));
+    await waitFor(() => expect(blobMock).toHaveBeenCalled());
   });
 
   it('should not download files when job is failed', async () => {
-    const linkMock = document.createElement('a');
-
     jest.clearAllTimers();
 
     useOkapiKy.mockClear().mockReturnValue({
@@ -130,13 +126,10 @@ describe('useCirculationLogExport', () => {
       { wrapper },
     );
 
-    linkMock.dispatchEvent = jest.fn();
-    jest.spyOn(document, 'createElement').mockReturnValue(linkMock);
-
     await result.current.requestExport();
 
     jest.runAllTimers();
 
-    expect(linkMock.dispatchEvent).not.toHaveBeenCalled();
+    expect(blobMock).not.toHaveBeenCalled();
   });
 });